Gui: split selection support declarations
